resourcemonitor: rewrite and simplify
In this patch we revisit pkg/resourcemonitor.
We want to reduce the API surface as much as possible
in order to simplify the package and trim down the size.
The idea is to get rid of the intermediate representation,
with all the extra complexity this encompesses.

Along the way, we add the option to refresh the allocatable
resources before each polling cycle, instead of just once
at the initialization time.
