+++ go.mod
@@ -19,7 +19,6 @@ require (
 	k8s.io/api v0.22.0
 	k8s.io/apimachinery v0.22.0
 	k8s.io/client-go v0.22.0
-	k8s.io/klog/v2 v2.9.0
 	k8s.io/kubelet v0.22.0
 	k8s.io/kubernetes v1.22.0
 	sigs.k8s.io/yaml v1.2.0
